+++ src/main/java/org/eclipse/tractusx/selfdescriptionfactory/service/converter/RegCodeMapper.java
@@ -32,7 +32,8 @@ public class RegCodeMapper {
         this.prefix = prefix;
     }
     private static final Map<RegistrationNumberSchema.TypeEnum, String> regCodeMapper = Map.of(
-            RegistrationNumberSchema.TypeEnum.TAXID, "local",
+            RegistrationNumberSchema.TypeEnum.TAXID, "taxID",
+            RegistrationNumberSchema.TypeEnum.LOCAL, "local",
             RegistrationNumberSchema.TypeEnum.VATID, "vatID",
             RegistrationNumberSchema.TypeEnum.EUID, "EUID",
             RegistrationNumberSchema.TypeEnum.EORI, "EORI",