+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 [Unreleased]
-
+## Added
+- New changes for service offering flow type.
+- legal Participant scheam updated for local.
+### Fixed
+- Fixed issue #239
+- Fixed issue #240 
 
 [2.1.19] - 2025-02-24
 ## Added